Improve type structure
